update : `_is_weighted`, `_get_precision`, `_threshold_calc` function moved to `pyrgg` engine.
